requirements.txt
discord.py==2.3.2
Flask==3.1.2
Requests==2.32.5
hypercorn==0.14.3